requirements.txt
torch
torchvision
Pillow
tqdm
scikit-learn